+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Migrate `PatientsTabScreen` to a fragment
 - Migrate `ReportsScreen` to a fragment
 - Migrate `OverdueScreen` to a fragment
+- Add `DatePickerKeyFactory` for getting date picker key based on `Country`
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 05 Apr 2021] Migrate app to use ViewBinding
 - [In Progress: 08 Apr 2021] Add automatic performance profiling for the Room database queries